chore: remove legacy Autocomplete
