--- src/components/FilterBase/FilterBase.i18n/en.json
@@ -1,4 +1,0 @@
-{
-    "Search": "Search...",
-    "Suggestions": ""
-}
