+++ src/components/GoalSelect/GoalSelect.i18n/en.json
@@ -1,0 +1,3 @@
+{
+    "Suggestions": "Suggestions"
+}
