+++ src/components/GoalSelect/GoalSelect.i18n/ru.json
@@ -1,4 +1,3 @@
 {
-    "Search": "Поиск...",
     "Suggestions": "Предложения"
 }
